+++ src/app-layer-parser.c
@@ -56,7 +56,7 @@
 #include "app-layer-smtp.h"
 #include "app-layer-modbus.h"
 #include "app-layer-enip.h"
-#include "app-layer-dnp3.h"
+#include "app-layer/dnp3/app-layer-dnp3.h"
 #include "app-layer-nfs-tcp.h"
 #include "app-layer-nfs-udp.h"
 #include "app-layer-ntp.h"
